requirements.txt
base58==0.2.2
boto==2.39.0
boto3==1.2.3
botocore==1.3.22
coverage==4.0.2
Django==1.8
docutils==0.12
filechunkio==1.6
future==0.15.2
futures==3.0.4
httpie==0.9.3
ipaddress==1.0.14
jmespath==0.9.0
mysql-connector-python-rf==2.1.3
pprofile==1.8
pygeoip==0.3.2
Pygments==2.1
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
six==1.10.0
slugify==0.0.1
tqdm==3.7.1
uguuAPI==1.0.0
wheel==0.24.0
wsgi-request-logger==0.4.4
djangorestframework==3.6.3